fix pm filter

+++ backend/create_marcas_pm_table.sql
@@ -13,7 +13,7 @@ CREATE INDEX idx_marcas_pm_usuario_id ON marcas_pm(usuario_id);
 
 -- Insertar todas las marcas existentes sin asignar
 INSERT INTO marcas_pm (marca, usuario_id)
-SELECT DISTINCT marca, NULL
+SELECT DISTINCT marca, NULL::INTEGER
 FROM productos_erp
 WHERE marca IS NOT NULL
 ON CONFLICT (marca) DO NOTHING;
